% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit.model}
\alias{fit.model}
\title{Fitting curve model and drawing samples}
\usage{
fit.model(
  stan.data,
  n.chain = 1,
  iter = 20000,
  warmup = 10000,
  cores = 1,
  initial.values = NULL,
  gen.inits = T,
  control.param = list(adapt_delta = 0.995, max_treedepth = 12, stepsize = 0.01),
  verbose = T,
  seed = NULL,
  ...
)
}
\arguments{
\item{stan.data}{An object of class VASCurves_Raw. See setup.data function.}

\item{n.chain}{The number of Markov chains to be run.}

\item{iter}{The number of iterations for sampling.}

\item{warmup}{The number of warm-up iterations to perform.}

\item{cores}{The number of cores available for running model.}

\item{initial.values}{Optional list of starting values for sampler.}

\item{gen.inits}{Logical value for whether initial values should be generated automatically using input VASCurves_Raw object. Only applies if initial.values is also NULL. See details.}

\item{control.param}{List of control settings for performing sampling.}

\item{verbose}{Would you like printed messages from sampler?}

\item{seed}{Seed for random number generation.}

\item{...}{Other parameters passed to rstan::sampling function.}
}
\value{
An object of class VASCurves_Res that includes the resultant stan object from the model fit, number of chains, VASCurves_Raw input object, diagnostics of the parameters traced, and the summary of the model fit. See details.
}
\description{
Fit the appropriate VAS curve model to the formatted data.
}
\details{
gen.inits generates initial values for each chain by initializing all individual and continuum effects to zero 
while simultaneously using a random multiplier between 0.9 and 1.1 applied to the prior specification for betameans and Bmeans.
Diagnostics in the VASCurves_Res object are calculated using the gelman.diag function in the coda package if n.chain > 1. Otherwise, the Rhat statistic from the rstan package is given for each parameter.
}
