% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priors.R
\name{get.priors}
\alias{get.priors}
\title{Generate prior values for model}
\usage{
get.priors(stan.data, seed = 123)
}
\arguments{
\item{stan.data}{A list of data created in the process of running setup.data.}

\item{seed}{Seed for random number generation.}
}
\value{
A named list of hyperparameter values for model.
}
\description{
Helper function used for calculating priors in setup.data function.
}
