% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_VASCurves_Res.R
\name{plot.VASCurves_Res}
\alias{plot.VASCurves_Res}
\title{Method for plot function in VASCurves}
\usage{
\method{plot}{VASCurves_Res}(object, subj_ids = "all", ...)
}
\arguments{
\item{object}{An object of class VASCurves_Res.}

\item{subj_ids}{Vector of string values detailing which subject ids to plot respective data. Defaults to "all" which gives a plot for all subjects.}

\item{...}{Other graphical parameters.}
}
\value{
Plots of the raw data for specified individuals in the input VASCurves_Res object as well as fitted curves.
}
\description{
Plotting method for an object of class VASCurves_Res. Plots the raw data for each individual in the dataset along with fitted curve(s).
}
