% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_VASCurves_Raw.R
\name{print.summary.VASCurves_Raw}
\alias{print.summary.VASCurves_Raw}
\title{Print method for summary function in VASCurves}
\usage{
\method{print}{summary.VASCurves_Raw}(object, ...)
}
\arguments{
\item{object}{An object of class VASCurves_Raw.}

\item{...}{Other parameters.}
}
\value{
An object of summary statistics of the data.
}
\description{
Method to print summary object returned from summary.VASCurves_Raw.
}
