% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_VASCurves_Res.R
\name{print.summary.VASCurves_Res}
\alias{print.summary.VASCurves_Res}
\title{Print method for summary function in VASCurves}
\usage{
\method{print}{summary.VASCurves_Res}(object, ...)
}
\arguments{
\item{object}{An object of class VASCurves_Res.}

\item{...}{Other parameters.}
}
\value{
An object of summary statistics of the data.
}
\description{
Method to print summary object returned from summary.VASCurves_Res.
}
