% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_data.R
\name{setup.data}
\alias{setup.data}
\title{Format given data for model fitting}
\usage{
setup.data(input.data, model.type, priors = NULL, seed = 123)
}
\arguments{
\item{input.data}{A data.frame containing data. Expecting columns Response (observed values), Time, and Subject. If model.type is Continuum, also expecting Continuum column.}

\item{priors}{A list of hyperparameter specifications needed for model fitting. Default option of NULL uses get.priors function to generate reasonable priors based on data. See details for more information.}

\item{seed}{Seed for random number generation.}

\item{model.name}{Character string detailing model to be fit. Current options are Continuum and No Continuum.}
}
\value{
An object of class VASCurves_Raw.
}
\description{
For a given set of data, format it appropriately into a list with parameters expected from model fitting functions.
}
\details{
For model.type = "No Continuum", priors is expecting a list with elements:
betameans (4-element vector for population curve parameters),
umeans (4-element vector for individual effects),
Bmeans (two element vector for log variance model),
usigmean (real value for individual effect for log variance),
betavar (4x4 covariance matrix for population parameters),
uvar (4x4 covariance matrix for individual curve parameters),
Bvar (2x2 covariance matrix for parameters of log variance model),
usigsd (positive value for variance of individual effect in log variance model).

For model.type = "Continuum", priors is expecting a list with elements:
betameans (4-element vector for population curve parameters),
cmeans (4-element vector for continuum effects),
umeans (4-element vector for individual effects),
Bmeans (two element vector for log variance model),
usigmean (real value for individual effect for log variance),
csigmean (real value for individual effect for log variance),
betavar (4x4 covariance matrix for population parameters),
cvar (4x4 covariance matrix for continuum parameters),
ucvar (4x4 covariance matrix for individual-continuum parameters),
uvar (4x4 covariance matrix for individual parameters),
Bvar (2x2 covariance matrix for parameters of log variance model),
usigsd (positive value for variance of individual effect in log variance model),
csigsd (positive value for variance of continuum effect in log variance model),
ucsigsd (positive value for variance of individual-continuum effect in log variance model)
}
\examples{
\dontrun{
### No Continuum Curve Model
raw_data <- setup.data(input.data, "No Continuum")
### Continuum Model
raw_data <- setup.data(input.data, "Continuum")
}
}
