% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_VASCurves_Res.R
\name{summary.VASCurves_Res}
\alias{summary.VASCurves_Res}
\title{Method for summary function in VASCurves}
\usage{
\method{summary}{VASCurves_Res}(object, digits = 3, ...)
}
\arguments{
\item{object}{An object of class VASCurves_Res.}

\item{digits}{Number of digits to print for numeric variables.}

\item{...}{Other parameters.}
}
\value{
An object of summary statistics of the data.
}
\description{
Summary method for an object of class VASCurves_Res.
}
